Move ICanProvide append-metadata interfaces to Cratis.Chronicle.Events
ICanProvideEventSourceId, ICanProvideEventStreamId and ICanProvideSubject
move from Cratis.Chronicle.Reactors into Cratis.Chronicle.Events, alongside
EventStreamIdAttribute. These contracts describe append metadata shared by
both reactor side-effects and commands, so a neutral namespace keeps them
usable from either side.
