+++ Source/Clients/DotNET/Events/ICanProvideEventSourceId.cs
@@ -1,9 +1,7 @@
 // Copyright (c) Cratis. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using Cratis.Chronicle.Events;
-
-namespace Cratis.Chronicle.Reactors;
+namespace Cratis.Chronicle.Events;
 
 /// <summary>
 /// Defines a reactor (or any object) that can provide the <see cref="EventSourceId"/> to use when
