+++ Source/Clients/DotNET/Events/ICanProvideSubject.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Cratis. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace Cratis.Chronicle.Reactors;
+namespace Cratis.Chronicle.Events;
 
 /// <summary>
 /// Defines a reactor that can provide the <see cref="Subject"/> to associate with side-effect events
